Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > > Patch Set 1:
> > > 
> > > Greek has made some progress in Pootle and is now at 59%. Can we wait until the final string synchronization to decide on it?
> > 
> > I just sent a message to haiku-i18n. If Greek gets above 60% before the 22nd of May, we will include it again.
> 
> I'm a native speaker of the Greek language and I've been putting some time and effort into getting the translation done in time.
> 
> I haven't joined the mailing list yet, but I'd like to ask for a clarification; Is the 22nd of May a final deadline that will take place alongside the new beta release of Haiku?

I would suggest you subscribe to the mailing list and introduce yourself there as soon as possible. Right now you are adding suggestions, which is nice, but it does not add to the translation count. After you introduce yourself, we can give you permissions to actually submit translations. Greek is not that far out, it is only 2% shy of the 60% treshold, but it was removed because it isn't maintained and therefore dropped under the 60%.

Patch-set: 1
Reviewer: Gerrit User 1000012 <1000012@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>