Update patch set 1
Patch Set 1:

The year is mandatory in the tag: URI scheme. It is used because if we ever let haiku-os.org be transferred to someone else, they could decide to generate tags using it as well. We can use any year for which we owned the domain name, or can prove that no one else owned it. This means we can probably use 2001, or 0.

Using another scheme would require registering it with the IANA, which is why I went with tag: which is freely usable. I would prefer that we keep this an URI so we can change our mind later (either if we register hpkr: with the IANA, decide to use UUIDs) and also because it allows other repository providers to pick their own scheme if they need to (maybe they will also go with tag: or maybe they will use something else).

Patch-set: 1
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>