Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Well, I'd prefer using just :// (no scheme) if we can, then?

That's not allowed by the RFC on URIs. I actually looked at the IANA registry for URIs and we have only these two choices that don't require registration: uuid: (which are not user readable) and tag: (which has a somewhat fixed format), or using an http: url as we did before.

Anything else would require registration somewhere: custom schemes, using name:... or make little sense (mailto:, gopher:, irc:, whatever you could think of).

Maybe data: could be considered since we can put arbitrary things in it. But tag: seems to be exactly what we need and ensures correct namespacing (we own the domain at a given time, we can organize ourselves however we want inside that domain,year pair) so it makes the process as simple as possible for anyone creating a repo. With data: we would have to maintain our own set of good practises and it would probably end up quite similar to tag: if we want to avoid conflicts.

In any case, we can change these at each new beta release since we create new repos. So if we go through the IANA registration for hpkr: we can then use that for beta3 (I don't see the point, and we would then have to specify how other repo providers are to format their URIs).

Patch-set: 1
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>