Icon-O-Matic: replace SVG parser with nanosvg
The parser is based on an old example from agg. NanoSVG is originally
based on the same code, but has lots of bugfixes and is actively
maintained. So it makes sense to use it.

This is incomplete, there are several regressions so far.

Nanosvg revision 25241c5a8f8451d41ab1b02ab2d865b01600d949

Change-Id: I38ff9aa4e1d403c41979ebe42f7b45d4500a870c