Update patch set 11
Patch Set 11:

(3 comments)

> Patch Set 9:
> 
> (4 comments)
> 
> I'm still unsure how well this will integrate with the existing protocol loop API, but I think I'm starting to think of something. I'm adding Waddlesplash and Leorize to this because it may be relevant for the GSoC project.
> 
> Currently the HTTP API has a single "BHttpRequest" class that manages a bit of everything:
> - The connection and socket to the server,
> - The thread where to run the request,
> - And the request itself
> 
> I think this should be split in different parts. There would be a BHttpConnection to handle a connection with a server. Connections could be, for example, stored in a pool, and a request would try to reuse an existing but idle connection from there, before creating a new one. We could also have a maximum number of connections per host, and if they are all busy, new requests would wait. Maybe the connection should run in its own thread, or maybe that should be split too, but we can see about that later.
> 
> The BHttpRequest class would then handle purely the "request" side of things. It doesn't need to bother about opening the socket or threading details. It just uses the existing connection object to manage that. We need to define which parts of the BHttpRequest code should run in the connection thread, and which should run in the caller thread (or if maybe this should all be synchronous in a single thread, leaving it up to the caller to set up threads, but I think that isn't great for ease of use of the API). I think it will solve some ownership problems, because currently a request deletes itself and stops its own thread when it's complete from the network side, but there may still be things "in flight" with the asynchronous notifications to the caller thread. As a result, the code for asynchronous notifications needs to archive and unarchive all relevant objects (BHttpResult, received data, ...) resulting in many unneeded copies. If we could better handle the lifetime of the BHttpRequest (the simplest way being that the caller is responsible for deleting it when its work is done, which would now be easier because the class is not associated with a thread).
> 
> This also matches better with the way FTP works, because FTP is really designed with the idea that you have a connection, and then you have requests, which are clearly independant of the connection.
> 
> So I guess we now need a BFtpRequest that uses this BFtpConnection, and handles that part? And then we should make it use an API similar to BHttpRequest, so that a client (a web browser, for example), can use either without too much problems (some things will need to be different, but the basic workflow and things like getting data should be kept identical).
> 
> I think HTTP2 will also be similar to FTP in the sense that there are multiple requests using the same connection, with the additional fact that in HTTP2, there may be multiple requests at the same time on a connection (the protocol itself being more asynchronous than HTTP 1).

I agree. I just didn't know how to keep a connection open with the current system. Should I make a class BFtpConnection that infinitely loops and constantly polls for any new requests? In my mind the whole system should be changed, I didn't want to work around it. Ideally we would multiplex everything just like HTTP/2.0 does. I think FTP is connection oriented in the sense that its requests are more like simple commands. So we shouldn't separate the concerns here? Or we could according to the new system, whee I think the connections and requests should communicate with each other through pipes which can be poll()ed alongside the connections sockets.

I think it should work like this:
A server would listen to new connection through a pipe which will be poll()ed. When poll() returns a new socket will be opened and saved by the server. A request would then be created with a pipe between it and the server. When the request needs to share data with the server (as in download or upload a file in FTP) it should just put in the pipe. This pipe will be poll()ed with the rest of the sockets. And as such the server should store information relating each socket(connection) with its requests and helps share data between them.

Patch-set: 11
Reviewer: Gerrit User 1000206 <1000206@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>