Update patch set 11
Patch Set 11:

> I agree. I just didn't know how to keep a connection open with the current system. 

Yes, that's indeed the main problem of the current API, it was not designed with that in mind. Attempting to implement FTP using it shows its limits.

> Should I make a class BFtpConnection that infinitely loops and constantly polls for any new requests? In my mind the whole system should be changed, I didn't want to work around it. Ideally we would multiplex everything just like HTTP/2.0 does. I think FTP is connection oriented in the sense that its requests are more like simple commands. So we shouldn't separate the concerns here? Or we could according to the new system, whee I think the connections and requests should communicate with each other through pipes which can be poll()ed alongside the connections sockets.

The difficulty is that the system as a whole is asynchronous on both sides:
- On the network side, we don't want to block on the socket when there is nothing happening. But we do want to react quickly when something happens.
- On the client side, the client is probably a GUI app and has other things to manage (windows events and other UI interactions)
Moreover, the two are incompatible, as on one side we are blocking on sockets and file descriptors, while on the other we are using a BMessage based event loop. The main problem lies in meshing the two together efficiently, and there seem to be no way except by having separate threads (and the communication between the two threads is still problematic).

But at the same time, we also want to handle the really simple case of "but I just want to download a file", which ideally wouldn't need spawning this many threads, event loops, pipes, etc to get things done.

Also, there doesn't need to be a matching between threads and C++ objects. Maybe we want to have many sockets handled by a single thread? Would this perform better?

> 
> I think it should work like this:
> A server would listen to new connection through a pipe which will be poll()ed. When poll() returns a new socket will be opened and saved by the server. A request would then be created with a pipe between it and the server. When the request needs to share data with the server (as in download or upload a file in FTP) it should just put in the pipe. This pipe will be poll()ed with the rest of the sockets. And as such the server should store information relating each socket(connection) with its requests and helps share data between them.

That's a lot of moving information back and forth, however: the data is created on the client side, then it is sent in the pipe, then read back on the other side by the connection thread, and then written again to the socket. Performance wouldn't be great.

The other direction would also need a different solution, because the client thread wouldn't be able to block on the pipe, it needs to be notified using BMessages instead (this is done using BUrlProtocolAsynchronousListener currently, which converts the callbacks into messages).

I would consider a design which avoids the pipe as much as possible:

- The connection class is responsible for establishing a connection and keeping it alive between requests
- The "protocol roster" handles creating requests (as it does already). When you use MakeRequest, it will try to locate an existing and idle connection, but if none is available, it will instead spawn a new one
- The request and the connection are then bound together. At this point, the request code can be run in the connection thread, and the connection is not available for other requests (FTP) or requests make sure to synchronize themselves to keep the connection in working state (HTTP2). Currently the request is implemented by running code on both the connection and the client thread, but most of the work is on the connection side, with the client being merely notified of events, however, this doesn't work perfectly because the client sometimes wants to act according to the events, so this should be reconsidered.

Patch-set: 11