Update patch set 11
Patch Set 11:

> Alright, but as it stands I don't know what to do next. Should I wait for the current GSoC project to finish and then integrate FTP with it? Because I have no idea how to achieve this with the current system. Or should I try to work around it? Or should I try to design a new system for FTP that doesn't use `libnetapi`?

It's not possible with the current system. So let's wait to see what Leorize has in mind for the GSoC project and what the planning would look like. There is no point to develop this outside of libnetapi, because the end goal is to get FTP support in the web browser, which uses libnetapi. But experimenting with this certainly helps to better show and understand the limits of the current design and think about the new one.

Patch-set: 11
CC: Gerrit User 1000011 <1000011@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>
CC: Gerrit User 1000035 <1000035@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>
Work-in-progress: false