Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Why add virtual? This classes are not intended to be extended.

The methods are already virtual from the parent class. If we were allowed to use C++11 I would mark them final or override, but currently we can't do that because gcc2. It's just an indicator to developers that this is meant to implement the superclass method.

Patch-set: 1
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>