Update patch set 4
Patch Set 4:

Hi; I have change this around a bit to alter the `url` field to `identifier` and hence introduce a version 3 (hopefully the last).  I have also introduced here the swapping of the well-known legacy to new "identifier" for the version 1 files which is from the era when there was no separation between the base-url and the identifier.

Patch-set: 4
Reviewer: Gerrit User 1000018 <1000018@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>