Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> IMHO, we should probably rename AutoDeleter at this point to UniquePointer (or something like that, as it has different semantics vs. std::unique_ptr), and add operator==/operator!= for easy comparison without Get().

AutoDeleter does not ensure uniqueness, so that wouldn't be a great choice of name. OwningPtr may make sense, but I think AutoDeleter is clear enough and matches with AutoLock which is used in a similar way.

I'm not sure about == and !=, doesn't it have an operator bool() to easily check for NULLness already? So you can simply do if(somePointer) ?

Patch-set: 1
CC: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>
Work-in-progress: false