Update patch set 1
Patch Set 1: Code-Review-1

I'm not a huge fan of this, as the ownership is usually clear enough (a good example of this would be ServerWindow, where there is pretty much a single delete in the whole file), and it makes the code more cumbersome to read. It's often not a huge win IMO, but rather annoying. The same knowledge could be put into a comment without clobbering the rest of the code.

In any case, I would prefer a new class for this, like "OwnedPointer" or something like that. Or why not simply use std::unique_ptr instead? This would even provide a few more advantages.

I would also prefer not to add TODOs like in those in ServerWindow. Please make sure it's needed or not, and then update the code accordingly.

Patch-set: 1
Reviewer: Gerrit User 1000016 <1000016@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>
Label: Code-Review=-1