kernel/locks: Remove ignore_unlock_count.
This was actually just plain wrong: if ignore_unlock_count was > 0
and the holding thread called unlock(), MUTEX_FLAG_RELEASED
would never be set and the lock would never be able to be acquired again.

As far as I can tell, there is no reason to ignore unlocks, ever;
if no threads are waiting, then mutex_unlock() will act appropriately.
So all we need to do is increment the lock's count here,
as we are relinquishing our request for locking.

On the other hand, if we did not find our structure in the lock,
that means we own the lock; so to return with an error from here
would also result in a deadlock, as nothing would ever set
MUTEX_FLAG_RELEASED.

The first of these problems would only occur under non-KDEBUG
(i.e. non-nightly) kernels, as ignore_unlock_count was only
set from there. But the latter could occur on any kernel.

May fix #16044, as it appears there is a case where ACPICA
calls this with a timeout of 0 (we should make this be
a mutex_trylock, anyway.)