Update patch set 3
Patch Set 3:

> I still don't think this is the right way to proceed. It is almost adding double the amount of code it removes, to fix two small bugs? I would prefer we just fix the issues (one of them has a patch pending elsewhere already) and stick with the layouts.

Dynamic layout would make sense if it were valid to consider each "package view" (row) in isolation, but this is most definitely not the case here.  In this case the structure of the overall UI element is a quasi-table.  The ratings stars display should line-up and the package summary should line-up and the icons should line up etc...  The layouts are not independent but are being executed entirely independently.  Repeating the same layout engine execution for each and every row makes no sense to me at all.

Furthermore the layout comes at a cost.  It is not executed once and then reused; it is instead executed for every row.  This means that if the application brings up 1000 or 10000 packages then every row has to go through automated layout even though the structure of the view is essentially static.  This is a large computational cost that yields nothing.

I cannot see a case here for dynamic layout.

The main class which has changed is `FeaturedPackagesView.cpp`.  The difference in code size for this class does not seem to represent a doubling and I find the new code easier to understand.  Furthermore the new code also supports keyboard navigation and contains some helpful code-comments which may explain the additional 206 lines of code (incl whitespace).

I am unconvinced of the argument around not making this change.

Patch-set: 3
Reviewer: Gerrit User 1000018 <1000018@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>