Update patch set 3
Patch Set 3:

Axel, I'm removing your -2 review since the issue has been adressed. Remember that -2 remains when a patch is updated and normally blocks merging changes, only yourself or a Gerrit admin can bypass it. For problems that only require a rework of the patch, you can use -1 which does not block merging when the patch is updated.

If this does not seem suitable we can discuss it on the mailing list and change the Gerrit configuration (for example the -2 can be removed automatically when a new version of a change is updated).

Patch-set: 3
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>