nvme_disk: Fix nvme_qpair_submit_queued_requests.
I added this function to make the queued requests logic more robust,
but I failed to notice that since it is called from nvme_qpair_submit_request,
it would just result in a deadlock of any qpair that had requests
queued.

May help with some of the "qpair lockup" tickets.