Update patch set 1
Patch Set 1:

The logic behind this change is the following;

The menu and the affected options are supposed to offer quick access to certain things, so that the user doesn't manually look in their system for them; Applications, Folders, Documents. So, I readjusted the order of those options depending on what's going to be more cluttered and less cluttered in a real-world scenario.

I'm not sure how I can articulate and convey the logic behind my change properly, but I'll try my best-- Under normal circumstances, I believe that the user uses a smaller amount of applications compared to folders to an extent, a smaller amount of folders compared to documents. I think that the amount of objects that fall under each category is relevant to the usefulness of each submenu, considering that the usefulness of the menu is directly dependent on its volatility. I think that if it is less likely for the user to find what they're looking for with a specific feature that's meant to be convenient but changes all the time, the user will eventually learn to do it manually.

I believe that putting the most useful 'options' in this case on top makes sense, considering that the people who go further down are very likely to be only interested in the applications. Putting more useful options in between of less used options and the applications (which are irrelevant) decreases the user experience.

Patch-set: 1
Reviewer: Gerrit User 1000173 <1000173@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>