Update patch set 1
Patch Set 1:

Testing your patch, I just had a look at what appears first in HaikuDepot on launch.

The package "0 A.D." is correctly greying out the Changelog and Contents tab labels. The "0ad_data" package below it only greys out the Contents label.

Would it be easily possible to show a disabled "Contents" tab with the same background grey with a "The contents of the package is only available once it's installed." text, similar to a disabled changelog?

Patch-set: 1
CC: Gerrit User 1000004 <1000004@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>