Update patch set 1
Patch Set 1:

> Isn't the order of these "Recent *" items quite arbitrary?
> 
> It has been in this order for a long time, so changes should IMO give substatial advantages. 

I mean, it's a small change that just improves the organization and the user experience, so of course, the change isn't going to offer substantial advantages, since I didn't opt for a full refactor. But even if I did, what would be the difference? Refactoring the code (I presume that this is one of the few scenarios which would fall under the list of things that would offer 'substantial advantages') would just make the concern regarding the people that are already used to what they have right now would be even greater. I believe that there's a potential for people being slightly confused while learning how to use the UI and that there's also a potential for accidental misclicks, rather than just saving a few pixel movements. Even if the latter is the only advantage from an objective standpoint, why not take that advantage instead of leaving things the way they are right now and being conservative with things that could make sense instead but don't?

I feel like the order something that could be improved on and I doubt that the time that it has been left untouched for negates its usefulness. I'd even think the fact that the order has been like this for quite a bit now is even more of a reason to make that small change. I also don't believe that arbitrarities and the UI go well together. If not now, then when?

> I, for example, almost never use "Recent applications" but use QuickLaunch instead. I do use "Recent folders" though (would do more often, if it worked on non-boot volumes...) and also "Recent documents" from time to time.

Yeah, that's fair enough, but I think that it's safe to assume that QuickLaunch would be more likely to be used by power users or people that are well-adjusted to Haiku, rather than the targeted/intended majority. QuickLaunch (or any other launcher) isn't implemented as a part of the core operating system experience (as in 'by default', or mentioned in the 'User Guide'-- the only extensive mention of it on the website is from a blog post from the 22nd of January, 2017, according to some quick web searches) and it's something that's completely separate from the Deskbar. I think that improvements to a specific, commonly used functions are by all means better than redirections to other software, or expectations such as "the user could simply drag and drop the icons to the Desktop or organize their documents/folders better" (which effectively puts the need for the submenus up for debate)-- that just sets a precedent of expectations for the end user.

Patch-set: 1
Reviewer: Gerrit User 1000173 <1000173@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>