Update patch set 1
Patch Set 1:

> Do we actually need to bypass the cache here? I.e., if there is an entry in the cache, why not use it, then?

IIRC the cache needs a write lock and things get super complicated to manage all the locks (I tried, and found many ways to freeze the app_server). The glyph from the fallback font will be cached in the font that's missing it and got us here anyway, so this happens just once and then the glyph is stored in the cache at the first level and this code is not called again.

Patch-set: 1
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>