Update patch set 4
Patch Set 4: -Code-Review

Fixed the most obvious issues. Radial gradient still not perfect but when exporting to SVG and re-importing in old Icon-O-Matic the gradients already changes, so it's hard for me to test this properly.

I think we can merge as results are still much better than with the old code, and we can improve on this later as problems are found.

Patch-set: 4
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>
Label: Code-Review=0