Update patch set 1
Patch Set 1:

> I'm not sure that would allow the same level of performance. Latency is important for input devices, so minimizing the number of system calls is a good idea.

What's the difference here? The data has to make it to userland one way or another, so why does it matter for a specialized driver vs. the usb_raw one?

It will be 2-3x as many syscalls at most; we already have to have a few already, and syscall overhead is measured in microseconds (or less, actually.)

Patch-set: 1