Update patch set 2
Patch Set 2:

> Patch Set 2: Code-Review-1
> 
> I think it would be better if we added a new Jam BuildScript type to create these indexes, the same way files, directories, etc. are added to the image. That way this will be configurable from Jam based on profile or UserBuildConfig, and so we avoid hard-coding things in this file.

The list of indexed attributes does not change as much as the list of files. I don't think it is worth the extra complexity.

I understand the need to do that for files, as they are collected from several places in the buildsystem and are highly configurable. But the indexed attributes will be always the same.

In any case, this fixes the initial problem, if you think the list should be configurable we can make a new ticket about it. But having queries work out of the box seems quite important to me.

Patch-set: 2
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>