boot_loader: Add unusable memory to the "ignored" memory parameter.
This way, AboutSystem and others will actually display the true
amount of reserved memory.

I've also removed ACPI Reclaim memory from the "ignored" set,
as this memory is actually used and is not ignored/inaccessible.
It isn't returned to the system for use once the ACPI driver
is done with it, but that's a different matter.

Fixes #16056.