HaikuDepot: Refresh Repo Cache
When HD first loads in a new install, it will experience
a near-empty set of repositories.  This is because only
build-time repo caches are present.  This change will
cause HD to pull data from remotes on first load, in the
case that the cache is somehow missing, on a user-
initiated update and on start every 5 days.

The mechanism for tracking when tasks were performed is
generalized so that the mechanism can be used in other
areas of functionality later.

Resolves #14840