Update patch set 1
Patch Set 1:

The cache file is abstracted through API so I'm not sure if I am able to see how old the cache is.  Even if the file were accessible directly as a file; if the cache is older than 5 days then it will not be possible after that to check it every subsequent 5 days because it will check every start-time for as long as there are no cache changes.  This is bad for an infrequently updated repository.

Keeping a single totem file won't work because there are multiple repositories and users may add or remove repositories all the time.  For this reason there needs to be some record of when  each individual repository's cache was refreshed.

So I started to build this mechanism into the repository cache refreshing logic, but then realized that the class was getting too large and complex.  For this reason I have moved it into its own class.  From there it would then be easy to also allow other processes to only run from time to time as well.  By adding the {{{type}}} instance variable to the class it will be possible to segregate clients of the {{{LastDoneStore}}} service.

I think this seems reasonable and I can't see how it could be less complex.

Patch-set: 1
Reviewer: Gerrit User 1000018 <1000018@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>