Update patch set 1
Patch Set 1:

Some old BeOS programs use archived BPictures (stored in application resources) or otherwise manipulate picture data. So we need to provide a compatible format for them.

But this format has limitations and can't be extended as we want for new drawing operations, so a new and easier to manage format was designed for that.

I think in particular the callback table passed to BPicture::Play needs some care, fortunately we get passed the size of the table so we can know if newer callbacks are not available. I think the conversion code in this case allows to emulate some operations using only the old callbacks?

Patch-set: 1