Character Map: Added larger character popup on left-click
Fixes #15983.

Added left-click support to the existing function that pops up the context
menu for copying characters.  Instead of menu options, put a large version
of the currently-selected character in the popup on left-click.  The popup
character is 2.5 times as large as the current font size, because that
seemed nicer than 2 times.  A future enhancement could be to set a minimum,
maximum, or fixed size for the popup, but 2.5x seemed to work well for all
font sizes in my testing, so maybe that's not worth the effort.

Right-click functionality is unchanged.

Change-Id: I3b64fff3cafc02e2c54e34db6e58461b4e08b3b9