Update patch set 1
Patch Set 1:

> Can it break code that use vregs?

It depends on how it is used. Normal use in POSIX is just to save and restore the state. Some apps however try to alter the state in various ways before restoring it, or do read from in (this is the case in webkit for example).

Do we need a 32bit version of the changes too, or is this only for 64bit arch?

Patch-set: 1
CC: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>
Work-in-progress: false